requirements.txt
spidev==3.4
RPi.GPIO==0.7.0
pandas==1.0.3
matplotlib==3.2.1
PyWavelets==1.1.1